requirements.txt
Keras-Preprocessing==1.1.2
matplotlib==3.4.3
numpy==1.19.5
pandas==1.3.4
python-speech-features==0.6
PyYAML==6.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
tensorflow-gpu==2.6.0

